my-hinge-app/backend/requirements.txt
flask
flask-cors
requests
openai
spacy
spacytextblob
python-dotenv 